Finally a got working ic10emu export! (Thanks RyRy!)
